Move Python 3.10 compatibility into compat module.
